Arreglo de errores en el merge de Develop-transferencias a develop
