+++ src/main/kotlin/mx/com/team9/activities/Ingreso.kt
@@ -2,18 +2,14 @@ package mx.com.team9.activities
 
 import mx.com.team9.models.Categoria
 
-/**
- * Clase que representa un retiro de una cuenta
- */
-
-class Gasto(
+class Ingreso(
     override val monto: Double,
     override val descripcion: String,
     override val categoria: Categoria,
     override val cuentaOrigen: Cuenta
 ) : Movimiento() {
     init {
-        cuentaOrigen.saldo -= monto
+        cuentaOrigen.saldo += monto
         logTransaccion()
     }
 }
