+++ src/main/kotlin/mx/com/team9/activities/Usuario.kt
@@ -1,46 +1,65 @@
 package mx.com.team9.activities
 
+import java.util.UUID
+
 const val PASSWORD_LENGHT = 10
-class User(name: String, email: String, password: String) {
+class Usuario(
+    idUsuario: String,
+    nombre: String,
+    correo: String,
+    contrasena: String,
+    var cuentas: MutableList<Cuenta> = mutableListOf()) {
     /**
      * Clase que representa a un usuario
-     * @param name Nombre del usuario
-     * @param password Contraseña del usuario
+     * @param nombre Nombre del usuario
+     * @param correo Correo del usuario
+     * @param contrasena Contraseña del usuario
      */
 
+    // Constructor vacío
+    constructor() : this("", "", "", "")
     init {
-        println("El usuario $name ha ingresado")
+        println("El usuario $nombre ha ingresado")
     }
 
-    private var name = name
+    private var nombre = nombre
         set(value) {
             if (value.length >= 3) field = value
             else println("El nombre debe contener al menos 3 digitos")
         }
-    private var email: String = email
+    private var correo: String = correo
         set(value) {
             if (value.contains("@")) field = value
             else println("El email debe contener un @")
         }
 
-    private var password: String = password
+    private var contrasena: String = contrasena
         set(value) {
             if (value.length >= PASSWORD_LENGHT) field = value
             else println("El password debe contener al menos 10 digitos")
         }
 
     fun getName(): String {
-        return this.name
+        return this.nombre
     }
 
     fun getEmail(): String {
-        return this.email
+        return this.correo
     }
 
-    fun validatePassword(password: String): Boolean {
-        return this.password == password
+    fun validatePassword(contrasena: String): Boolean {
+        return this.contrasena == contrasena
     }
 
+    // Método para crear un usuario
+    fun crearUsuario(nombre: String, correo: String, contrasena: String) {
+        val id = UUID.randomUUID().toString()
+        val usuario = Usuario(id, nombre, correo, contrasena)
+    }
 
+    // Metodo para iniciar sesión
+    fun iniciarSesion(correo: String, contrasena: String): Boolean {
+        return this.correo == correo && this.contrasena == contrasena
+    }
 
 }
